Separate common_metadata and provider tests
